pyproject.toml
[project]
name = "mcp-servers"
version = "0.1.0"
description = "Personal MCP servers collection"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.0",
    "google-generativeai>=0.8.0",
    "Pillow>=10.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pillow-heif>=0.13.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["domains", "gemini_alt", "tablica", "exif_extractor", "plate_recognition"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
